+++ CHANGELOG.md
@@ -95,6 +95,7 @@ Date format: (YYYY-MM-DD)
 * Admin shopkeepers normalize empty trade permissions in the save data to null now, similar to AdminShopkeeper#setTradePermission.
 * Removed an old update mechanism for sign shops: If the sign block is not available at the time it is meant to be updated, this would update the sign later, once it is available again. This mechanism should no longer be required: The sign is already updated whenever it is spawned, and we dynamically spawn and despawn the sign with chunk loads, and dynamically respawn it if it is detected to be missing.
 * ItemData caches the meta type of items now instead of always determining it freshly from a newly serialized ItemMeta instance.
+* The shop object types for the various mob shop objects use composition now instead of deriving their own shop object type from a base class.
 * Various other minor internal refactors and Javadoc improvements.
 
 **Message changes:**